requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
python-dotenv==1.0.1
pydantic==2.10.3
psycopg2-binary==2.9.9
